+++ Tests/Functional/Controller/Backend/ErrorHandlingTest.php
@@ -123,6 +123,7 @@ final class ErrorHandlingTest extends AbstractFunctionalTestCase
             $pageRenderer,
             $backendUriBuilder,
             $extensionConfiguration,
+            new \Psr\Log\NullLogger(),
         );
     }